reorganised the page
