+++ src/components/language/Language.module.css
@@ -40,3 +40,16 @@
   flex-direction: row;
   gap: 0.5rem;
 }
+
+.language-selector {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+  & p {
+    display: inline-block;
+    font-size: 0.75rem;
+    font-style: normal;
+    font-weight: 600;
+    color: var(--black);
+  }
+}
