+++ src/components/language/LanguageList.tsx
@@ -4,14 +4,21 @@ import style from "./Language.module.css";
 
 interface LanguageListProps {
   languages: Array<string>;
+  showTitle?: boolean;
 }
 
-const LanguageList: React.FC<LanguageListProps> = ({ languages }) => {
+const LanguageList: React.FC<LanguageListProps> = ({
+  languages,
+  showTitle = false,
+}) => {
   return (
-    <div className={style["language-badge-container"]}>
-      {languages.map((lang) => (
-        <Language lang={lang} />
-      ))}
+    <div className={style["language-selector"]}>
+      {showTitle && <p>Language</p>}
+      <div className={style["language-badge-container"]}>
+        {languages.map((lang) => (
+          <Language lang={lang} />
+        ))}
+      </div>
     </div>
   );
 };
