--- src/pages/components/LanguageSelector.module.css
@@ -1,47 +1,0 @@
-.language-selector {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-  & p {
-    display: inline-block;
-    font-size: 0.75rem;
-    font-style: normal;
-    font-weight: 600;
-    color: var(--black);
-  }
-}
-
-.language-badge-container {
-  display: flex;
-  flex-direction: row;
-  gap: 0.5rem;
-}
-
-.language-indicator {
-  display: inline-flex;
-  border-radius: 0.25rem;
-  border: 1px dashed var(--black);
-  max-width: 1.5rem;
-  justify-content: center;
-  align-content: center;
-
-  font-size: 0.625rem;
-  font-style: normal;
-  font-weight: 600;
-  letter-spacing: -0.01563rem;
-  padding: 0.2rem;
-  transition: all 250ms cubic-bezier(0.4, 0, 0.2, 1);
-}
-
-.language-indicator:active {
-  border-radius: 0.25rem;
-  border: 2px solid var(--black);
-  background: #fff;
-  font-style: normal;
-  font-weight: 600;
-  letter-spacing: -0.01563rem;
-  padding: 0.2rem;
-  transform: translateY(-0.5rem);
-  transition: all 250ms cubic-bezier(0.4, 0, 0.2, 1);
-  box-shadow: 0px 6px 2px 0px rgba(0, 0, 0, 0.05);
-}
