+++ src/pages/main/components/sidebar/Sidebar.module.css
